+++ layers/GatoGraphQLForWP/plugins/gatographql/CHANGELOG.md
@@ -11,6 +11,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Added fields `GenericCustomPost.update`, `Root.updateCustomPost` and `Root.createCustomPost` (#2663)
 - Added documentation for integration with Polylang (#2664)
 - Added module type "Integrations" (#2665)
+- Return an EnumString type on `GenericCategory.taxonomy` and `GenericTag.taxonomy` (#2666)
 
 ### Fixed